7thsem/requirements.txt
# Socio-Technical Service Operations System
# Version-pinned for reproducibility

# Core dependencies
torch>=2.0
kafka-python>=2.0
simpy>=4.0
streamlit>=1.30
plotly>=5.18
numpy>=1.24
pandas>=2.0

# Research-grade additions
scikit-learn>=1.3.0  # For Gaussian Process
scipy>=1.11.0  # For statistical tests
seaborn>=0.12.0  # For publication-quality plots
matplotlib>=3.7.0  # For visualizations
